move : BuyRepository 분리
