+++ src/main/java/bc1/gream/domain/buy/repository/BuyRepositoryCustom.java
@@ -1,6 +1,6 @@
-package bc1.gream.domain.order.repository;
+package bc1.gream.domain.buy.repository;
 
-import bc1.gream.domain.order.entity.Buy;
+import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.product.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
